Add files via upload
Contains selection sort, quick sort, and insertion sort and the main which creates an array list for each of the sorts and then perfors operations with each sort such as returning the middle element, Searches for the number 10 & returns its index, -1 otherwise, Inserts the number 230 at the first position of the array, Deletes the second element (or first if the second doesn’t exist), Sorts the
Sorts the array using a recursive mostly < O(n2) sort like quicksort or mergesort. Then it prints the running times for each operation/sort.